+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add UnwrapFrame function to extract a single frame from an error.
+  You can use this to implement your own trace formatting logic.
+
 ### Fixed
 - cmd/errtrace: Don't exit with a non-zero status when `-h` is used.
 - cmd/errtrace: Don't panic on imbalanced assignments inside defer blocks.